# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Calendros API Documentation",
        "description": "API documentation for Calendros - A comprehensive calendar management system",
        "contact": {
            "email": "support@calendros.example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://http://calendros.test/api",
            "description": "Local Development Server"
        },
        {
            "url": "https://prm.positive.io.vn/api",
            "description": "Production Server"
        },
        {
            "url": "https://calendros.montserrat.id.vn/api",
            "description": "Test Server"
        }
    ],
    "paths": {
...
        "/events/{event_id}/invites": {
            "get": {
                "tags": [
                    "Invites"
                ],
                "summary": "List event invites",
                "description": "Get all invites for an event (event owner only)",
                "operationId": "1d4fa4512e45858c45664dadde6ea5b9",
                "parameters": [
                    {
                        "name": "event_id",
                        "in": "path",
                        "description": "Event ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of invites",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "invitee_email": {
                                                        "type": "string",
                                                        "example": "invitee@example.com"
                                                    },
                                                    "status": {
                                                        "type": "string",
                                                        "enum": [
                                                            "pending",
                                                            "accepted",
                                                            "rejected"
                                                        ],
                                                        "example": "pending"
                                                    },
                                                    "user_id": {
                                                        "type": "integer",
                                                        "example": 2
                                                    },
                                                    "event_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden - Not event owner"
                    },
                    "404": {
                        "description": "Event not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Invites"
                ],
                "summary": "Invite user to event",
                "description": "Send an invitation to a user for an event (event owner only). Queues an email notification.",
                "operationId": "56c7fc937542cc16fa6eb4467c7a4da7",
                "parameters": [
                    {
                        "name": "event_id",
                        "in": "path",
                        "description": "Event ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "invitee_email"
                                ],
                                "properties": {
                                    "invitee_email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "colleague@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Invitation sent successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 3
                                                },
                                                "invitee_email": {
                                                    "type": "string",
                                                    "example": "colleague@example.com"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "pending"
                                                },
                                                "event": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/invites/me": {
            "get": {
                "tags": [
                    "Invites"
                ],
                "summary": "List my invites",
                "description": "Get all event invites for the authenticated user",
                "operationId": "badaee38602ff40316b290293ceaadc9",
                "responses": {
                    "200": {
                        "description": "List of user's invites",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "invitee_email": {
                                                        "type": "string",
                                                        "example": "me@example.com"
                                                    },
                                                    "status": {
                                                        "type": "string",
                                                        "example": "pending"
                                                    },
                                                    "event": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "title": {
                                                                "type": "string",
                                                                "example": "Team Meeting"
                                                            },
                                                            "start_time": {
                                                                "type": "string",
                                                                "format": "datetime"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/invites/{id}/accept": {
            "post": {
                "tags": [
                    "Invites"
                ],
                "summary": "Accept invitation",
                "description": "Accept an event invitation. Only the invitee can accept.",
                "operationId": "be4bd612e5c0e5bb7f334160be813a0a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Invite ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Invitation accepted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "accepted"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden - Not your invite"
                    },
                    "404": {
                        "description": "Invite not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/invites/{id}/reject": {
            "post": {
                "tags": [
                    "Invites"
                ],
                "summary": "Reject invitation",
                "description": "Reject an event invitation. Only the invitee can reject.",
                "operationId": "ca11095dae7acefc939b154bae0f7847",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Invite ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Invitation rejected",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "rejected"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden - Not your invite"
                    },
                    "404": {
                        "description": "Invite not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/invites/{id}": {
            "delete": {
                "tags": [
                    "Invites"
                ],
                "summary": "Delete invitation",
                "description": "Delete an event invitation (event owner only)",
                "operationId": "6b26230945cd9e874c489a4d2021c499",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Invite ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Invitation deleted successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden - Not event owner"
                    },
                    "404": {
                        "description": "Invite not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter your Sanctum token",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}